Mouse look upgrade
Before, I inherited terrible mouse movement from SDL1.2.  It was based on an invisible cursor that when hit the edge of the screen, would reset the position..  awful!

Now we use SDL_SetRelativeMouseMode and mouse states.  It is soooooo much smoother!  As it should be for any discerning Quake player!