Removal of PCX Screenshots
This is a modern move, just so folks can actually screenshot.  I've made it retro though!

It uses the SDL Renderer and buffers to get the fullsize scaled image so that it appears as large as it does on your monitor and is not captured at 320x200.  Super chad move!

Screenshots also now save in the same directory as the binary, so they don't pollute the 'id1' folder.

Also uses std_image_write.h - one of my favourite headers to use.  Thanks!!!